Created form to edit title.
